Properly handle inter-routine comms on 'TimedTryLock()'.
This new pattern uses one channel for inter-routine communication.
It is tricky to execute, but the pattern makes sense. Basically,
close the "signals" channel if the timeout gets hit, or if we want
to unlock the Mutex. In the background routine, first check if the
channel is open. If it is not, then Unlock() the Mutex and bail.
If the channel is open, tell the main routine we locked the Mutex,
and then wait to see if the "flock" is successful. If the channel
gets closed, then the main routine failed.

Also, use 'time.NewTimer()' so we can stop it and let the GC
take care of it (per the func's doc).