Removed test harness' 'once' arg. Refactored test harness.
A dead lock was possible when an error was thrown by one of the
test harness' go routines. When the routine tried to send the error,
the main thread was never able to receive it because it was waiting
on the sync.WaitGroup. Since the routine calls the 'WaitGroup.Done()'
method after sending to the channel, there was no way for the main
thread to continue executing.